+++ src/vec3.rs
@@ -3,13 +3,13 @@ use std::ops;
 use crate::random_between;
 
 #[derive(Default, Debug, PartialEq, Clone, Copy)] // TODO: is Copy ok?
-pub struct Vec3(f32, f32, f32);
+pub struct Vec3(f64, f64, f64);
 
 pub type Point3 = Vec3; // TODO: arithmetic safety (no point + point)
 pub type Color = Vec3; // TODO: omg this is too dangerous
 
 impl Vec3 {
-    pub fn new(x: f32, y: f32, z: f32) -> Self {
+    pub fn new(x: f64, y: f64, z: f64) -> Self {
         Self(x, y, z)
     }
 
@@ -25,27 +25,27 @@ impl Vec3 {
         }
     }
 
-    pub fn x(&self) -> f32 {
+    pub fn x(&self) -> f64 {
         self.0
     }
 
-    pub fn y(&self) -> f32 {
+    pub fn y(&self) -> f64 {
         self.1
     }
 
-    pub fn z(&self) -> f32 {
+    pub fn z(&self) -> f64 {
         self.2
     }
 
-    pub fn length(&self) -> f32 {
+    pub fn length(&self) -> f64 {
         self.length_squared().sqrt()
     }
 
-    pub fn length_squared(&self) -> f32 {
+    pub fn length_squared(&self) -> f64 {
         self.0 * self.0 + self.1 * self.1 + self.2 * self.2
     }
 
-    pub fn dot(&self, other: &Self) -> f32 {
+    pub fn dot(&self, other: &Self) -> f64 {
         self.0 * other.0 + self.1 * other.1 + self.2 * other.2
     }
 
@@ -63,7 +63,7 @@ impl Vec3 {
 
     // TODO: this function should only work on color
     pub fn color_to_bytes(self, samples_per_pixel: u32) -> [u8; 3] {
-        let scale = 1.0 / samples_per_pixel as f32;
+        let scale = 1.0 / samples_per_pixel as f64;
         let r = (self.0 * scale).clamp(0.0, 0.999) * 256.0;
         let g = (self.1 * scale).clamp(0.0, 0.999) * 256.0;
         let b = (self.2 * scale).clamp(0.0, 0.999) * 256.0;
@@ -81,7 +81,7 @@ impl ops::Neg for Vec3 {
 }
 
 impl ops::Index<usize> for Vec3 {
-    type Output = f32;
+    type Output = f64;
 
     fn index(&self, index: usize) -> &Self::Output {
         match index {
@@ -125,15 +125,15 @@ impl ops::Mul for Vec3 {
     }
 }
 
-impl ops::Mul<f32> for Vec3 {
+impl ops::Mul<f64> for Vec3 {
     type Output = Self;
 
-    fn mul(self, rhs: f32) -> Self::Output {
+    fn mul(self, rhs: f64) -> Self::Output {
         Self(self.0 * rhs, self.1 * rhs, self.2 * rhs)
     }
 }
 
-impl ops::Mul<Vec3> for f32 {
+impl ops::Mul<Vec3> for f64 {
     type Output = Vec3;
 
     fn mul(self, rhs: Vec3) -> Self::Output {
@@ -141,24 +141,24 @@ impl ops::Mul<Vec3> for f32 {
     }
 }
 
-impl ops::MulAssign<f32> for Vec3 {
-    fn mul_assign(&mut self, rhs: f32) {
+impl ops::MulAssign<f64> for Vec3 {
+    fn mul_assign(&mut self, rhs: f64) {
         self.0 *= rhs;
         self.1 *= rhs;
         self.2 *= rhs;
     }
 }
 
-impl ops::Div<f32> for Vec3 {
+impl ops::Div<f64> for Vec3 {
     type Output = Self;
 
-    fn div(self, rhs: f32) -> Self::Output {
+    fn div(self, rhs: f64) -> Self::Output {
         (1.0 / rhs) * self
     }
 }
 
-impl ops::DivAssign<f32> for Vec3 {
-    fn div_assign(&mut self, rhs: f32) {
+impl ops::DivAssign<f64> for Vec3 {
+    fn div_assign(&mut self, rhs: f64) {
         self.0 /= rhs;
         self.1 /= rhs;
         self.2 /= rhs;
